+++ Tests/MapboxSearchIntegrationTests/SearchEngineIntegrationTests.swift
@@ -5,7 +5,11 @@ import CoreLocation
 class SearchEngineIntegrationTests: MockServerTestCase {
     
     let delegate = SearchEngineDelegateStub()
-    lazy var searchEngine = SearchEngine(accessToken: "access-token", locationProvider: DefaultLocationProvider())
+    lazy var searchEngine = SearchEngine(
+        accessToken: "access-token",
+        locationProvider: DefaultLocationProvider(),
+        supportSBS: true
+    )
     
     override func setUpWithError() throws {
         try super.setUpWithError()